requirements.txt
# Streamlit
streamlit>=1.37
python-dotenv>=1.0.1

# LangChain (同一世代で揃える)
langchain==0.2.11
langchain-core==0.2.29
langchain-community==0.2.10
langchain-openai==0.2.0
langchain-text-splitters==0.2.2

# OpenAI & tokenizer
openai>=1.35
tiktoken>=0.6

# Vector store
chromadb==0.5.3

# Loaders / parsing（WebBaseLoader用のHTMLパーサ揃え）
beautifulsoup4>=4.12
lxml>=5.2
html5lib>=1.1
requests>=2.31

# Data
pandas>=2.1

# SQLite(FTS5) パッチ（使うなら残す）
pysqlite3-binary==0.5.2.post3

# （任意）docxを使う場合のみ
# python-docx>=1.1.0

pymupdf>=1.24      # PyMuPDFLoader 用
docx2txt>=0.8      # Docx2txtLoader 用（python-docx とは別物）